refactor: use pyjwt instead of python-jose
